% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gmediateout.R
\name{summary.gmediateout}
\alias{summary.gmediateout}
\title{Gmediate Object Summaries}
\usage{
\method{summary}{gmediateout}(object, ...)
}
\arguments{
\item{object}{output object of gmediate function}

\item{...}{arguments passed to other methods}
}
\description{
This function generates result summary table of gmediate function output object
}
